Refactor student application wizard into modular steps
